member can enter an event
Members can now enter event by clicking on the enter button on an events card

+++ hoaxify/hoaxify/src/main/java/com/hoaxify/hoaxify/event/EntrantRepository.java
@@ -2,7 +2,6 @@ package com.hoaxify.hoaxify.event;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntrantRepository extends JpaRepository<Entrants, Integer> {
-
+public interface EntrantRepository extends JpaRepository<Entrants, Long> {
 
 }
